Quality Refactor & Improvements
Added a incremental UTF8 validator. now we don't need to wait the full message to arrive to validate it. we can validate it on the fly as it arrives. this comes with an important advantage. we can detect invalid UTF8 early. now we also have a hard limit on the maximum message size.